requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
black==21.11b1
celery==5.2.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.9
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-countries==7.2.1
django-filter==21.1
django-redis==5.1.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
importlib-metadata==4.8.2
kombu==5.2.2
Markdown==3.3.6
mypy-extensions==0.4.3
names==0.3.0
pathspec==0.9.0
platformdirs==2.4.0
prompt-toolkit==3.0.23
psycopg2==2.9.2
PyJWT==2.3.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==3.5.3
regex==2021.11.10
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==4.0.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.6.0
